Use unsigned long instead of unsigned int

+++ alobe/event.cpp
@@ -3,14 +3,14 @@
 /**
  * Constructor
  */
-Event::Event(unsigned int step)
+Event::Event(unsigned long step)
 {
 }
 
 /**
  * Trigger event action on an actor with a given step
  */
-void Event::action(Actor & actor, unsigned int step)
+void Event::action(Actor & actor, unsigned long step)
 {
     if (step == my_triggerStep) {
         this->filteredAction(actor);
