+++ alobe/periodicEvent.h
@@ -12,11 +12,11 @@
 class PeriodicEvent: public Event
 {
     public:
-        PeriodicEvent(unsigned int step = 0, unsigned int period = 1); 
-        void action(Actor & actor, unsigned int step);                 // Change event filter rules (add periodicity)
+        PeriodicEvent(unsigned long step = 0, unsigned long period = 1); 
+        void action(Actor & actor, unsigned long step);                  // Change event filter rules (add periodicity)
 
     private:
-        unsigned int my_triggerPeriod;
+        unsigned long my_triggerPeriod;
 };
 
 #endif // __ALOBE__PERIODIC_EVENT__
